+++ src/main/java/org/example/trab_dsweb/service/WorkerService.java
@@ -2,16 +2,21 @@ package org.example.trab_dsweb.service;
 
 import lombok.AllArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
+import org.example.trab_dsweb.dao.JobApplicationDAO;
 import org.example.trab_dsweb.dao.WorkerDAO;
 import org.example.trab_dsweb.dto.CreateWorkerDTO;
 import org.example.trab_dsweb.dto.ReturnWorkerDTO;
+import org.example.trab_dsweb.exception.exceptions.BadRequestException;
 import org.example.trab_dsweb.exception.exceptions.ConflictException;
 import org.example.trab_dsweb.exception.exceptions.NotFoundException;
 import org.example.trab_dsweb.model.Worker;
+import org.springframework.context.MessageSource;
+import org.springframework.context.i18n.LocaleContextHolder;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Locale;
 import java.util.UUID;
 import java.util.stream.StreamSupport;
 
@@ -19,14 +24,17 @@ import java.util.stream.StreamSupport;
 @Service
 @AllArgsConstructor
 public class WorkerService {
+    private final JobApplicationDAO jobApplicationDAO;
     private BCryptPasswordEncoder encoder;
     private final WorkerDAO workerDAO;
+    private final MessageSource messageSource;
+    private final Locale locale = LocaleContextHolder.getLocale();
 
     public ReturnWorkerDTO findWorkerById(UUID id) {
         Worker worker = workerDAO.findById(id)
                 .orElseThrow(() -> {
                     log.error("Worker not found with ID={}", id);
-                    return new NotFoundException("Worker not found");
+                    return new NotFoundException(messageSource.getMessage("error.worker.notfound", null, locale));
                 });
 
         return new ReturnWorkerDTO(
@@ -54,12 +62,12 @@ public class WorkerService {
     public void createWorker(CreateWorkerDTO data) {
         if (workerDAO.findByCpf(data.cpf()).isPresent()) {
             log.error("Worker with CPF={} already exists", data.cpf());
-            throw new ConflictException("Profissional com esse CPF já existe");
+            throw new ConflictException(messageSource.getMessage("error.worker.conflict.cpf", null, locale));
         }
 
         if (workerDAO.findByEmail(data.email()).isPresent()) {
             log.error("Worker with email={} already exists", data.email());
-            throw new ConflictException("Profissional com esse e-mail já existe");
+            throw new ConflictException(messageSource.getMessage("error.worker.conflict.email", null, locale));
         }
 
         Worker newWorker = new Worker();
@@ -77,14 +85,14 @@ public class WorkerService {
         Worker existingWorker = workerDAO.findById(id)
                 .orElseThrow(() -> {
                     log.error("Worker not found with ID={}", id);
-                    return new NotFoundException("Worker not found");
+                    return new NotFoundException(messageSource.getMessage("error.worker.notfound", null, locale));
                 });
 
         if (data.cpf() != null && !data.cpf().isEmpty()) {
             workerDAO.findByCpf(data.cpf()).ifPresent(worker -> {
                 if (!worker.getId().equals(id)) {
                     log.error("Duplicate CPF={} on update for Worker ID={}", data.cpf(), id);
-                    throw new ConflictException("Worker with this CPF already exists");
+                    throw new ConflictException(messageSource.getMessage("error.worker.conflict.cpf", null, locale));
                 }
             });
             existingWorker.setCpf(data.cpf());
@@ -94,7 +102,7 @@ public class WorkerService {
             workerDAO.findByEmail(data.email()).ifPresent(worker -> {
                 if (!worker.getId().equals(id)) {
                     log.error("Duplicate email={} on update for Worker ID={}", data.email(), id);
-                    throw new ConflictException("Worker with this email already exists");
+                    throw new ConflictException(messageSource.getMessage("error.worker.conflict.email", null, locale));
                 }
             });
             existingWorker.setEmail(data.email());
@@ -108,13 +116,25 @@ public class WorkerService {
             existingWorker.setName(data.name());
         }
 
+        if (data.birthDate() != null) {
+            existingWorker.setBirthDate(data.birthDate());
+        }
+
+        if (data.gender() != null) {
+            existingWorker.setGender(data.gender());
+        }
+
         workerDAO.save(existingWorker);
     }
 
     public void deleteWorkerById(UUID id) {
         if (!workerDAO.existsById(id)) {
             log.error("Attempt to delete non-existing Worker with ID={}", id);
-            throw new NotFoundException("Worker not found");
+            throw new NotFoundException(messageSource.getMessage("error.worker.notfound", null, locale));
+        }
+        if (!jobApplicationDAO.findAllByWorkerId(id).isEmpty()) {
+            log.error("Cannot delete Worker with ID={} because it has job applications", id);
+            throw new BadRequestException(messageSource.getMessage("error.worker.delete.conflict", null, locale));
         }
         workerDAO.deleteById(id);
     }
